chore: Move things around to improve DX and maintainability.
