+++ backend/src/slot-machine/helpers/slot-symbol.enum.ts
